feat: update game loop
